+++ packages/orq-rc/src/models/components/deployments.ts
@@ -235,7 +235,7 @@ export type FileT = {
   /**
    * The name of the file, used when passing the file to the model as a string.
    */
-  filename?: string | undefined;
+  filename: string;
 };
 
 export type Four = {
@@ -634,7 +634,7 @@ export type TwoFile = {
   /**
    * The name of the file, used when passing the file to the model as a string.
    */
-  filename?: string | undefined;
+  filename: string;
 };
 
 export type Two4 = {
@@ -2025,7 +2025,7 @@ export namespace Deployments2PrefixMessagesType$ {
 export const FileT$inboundSchema: z.ZodType<FileT, z.ZodTypeDef, unknown> = z
   .object({
     file_data: z.string(),
-    filename: z.string().optional(),
+    filename: z.string(),
   }).transform((v) => {
     return remap$(v, {
       "file_data": "fileData",
@@ -2035,7 +2035,7 @@ export const FileT$inboundSchema: z.ZodType<FileT, z.ZodTypeDef, unknown> = z
 /** @internal */
 export type FileT$Outbound = {
   file_data: string;
-  filename?: string | undefined;
+  filename: string;
 };
 
 /** @internal */
@@ -2045,7 +2045,7 @@ export const FileT$outboundSchema: z.ZodType<
   FileT
 > = z.object({
   fileData: z.string(),
-  filename: z.string().optional(),
+  filename: z.string(),
 }).transform((v) => {
   return remap$(v, {
     fileData: "file_data",
@@ -3697,7 +3697,7 @@ export namespace Deployments2Messages3Content4Type$ {
 export const TwoFile$inboundSchema: z.ZodType<TwoFile, z.ZodTypeDef, unknown> =
   z.object({
     file_data: z.string(),
-    filename: z.string().optional(),
+    filename: z.string(),
   }).transform((v) => {
     return remap$(v, {
       "file_data": "fileData",
@@ -3707,7 +3707,7 @@ export const TwoFile$inboundSchema: z.ZodType<TwoFile, z.ZodTypeDef, unknown> =
 /** @internal */
 export type TwoFile$Outbound = {
   file_data: string;
-  filename?: string | undefined;
+  filename: string;
 };
 
 /** @internal */
@@ -3717,7 +3717,7 @@ export const TwoFile$outboundSchema: z.ZodType<
   TwoFile
 > = z.object({
   fileData: z.string(),
-  filename: z.string().optional(),
+  filename: z.string(),
 }).transform((v) => {
   return remap$(v, {
     fileData: "file_data",