+++ packages/orq-rc/src/models/operations/createdatasetitem.ts
@@ -239,7 +239,7 @@ export type CreateDatasetItem2File = {
   /**
    * The name of the file, used when passing the file to the model as a string.
    */
-  filename?: string | undefined;
+  filename: string;
 };
 
 export type CreateDatasetItem24 = {
@@ -714,7 +714,7 @@ export type CreateDatasetItem2DatasetsFile = {
   /**
    * The name of the file, used when passing the file to the model as a string.
    */
-  filename?: string | undefined;
+  filename: string;
 };
 
 export type CreateDatasetItem2Datasets4 = {
@@ -1851,7 +1851,7 @@ export const CreateDatasetItem2File$inboundSchema: z.ZodType<
   unknown
 > = z.object({
   file_data: z.string(),
-  filename: z.string().optional(),
+  filename: z.string(),
 }).transform((v) => {
   return remap$(v, {
     "file_data": "fileData",
@@ -1861,7 +1861,7 @@ export const CreateDatasetItem2File$inboundSchema: z.ZodType<
 /** @internal */
 export type CreateDatasetItem2File$Outbound = {
   file_data: string;
-  filename?: string | undefined;
+  filename: string;
 };
 
 /** @internal */
@@ -1871,7 +1871,7 @@ export const CreateDatasetItem2File$outboundSchema: z.ZodType<
   CreateDatasetItem2File
 > = z.object({
   fileData: z.string(),
-  filename: z.string().optional(),
+  filename: z.string(),
 }).transform((v) => {
   return remap$(v, {
     fileData: "file_data",
@@ -4006,7 +4006,7 @@ export const CreateDatasetItem2DatasetsFile$inboundSchema: z.ZodType<
   unknown
 > = z.object({
   file_data: z.string(),
-  filename: z.string().optional(),
+  filename: z.string(),
 }).transform((v) => {
   return remap$(v, {
     "file_data": "fileData",
@@ -4016,7 +4016,7 @@ export const CreateDatasetItem2DatasetsFile$inboundSchema: z.ZodType<
 /** @internal */
 export type CreateDatasetItem2DatasetsFile$Outbound = {
   file_data: string;
-  filename?: string | undefined;
+  filename: string;
 };
 
 /** @internal */
@@ -4026,7 +4026,7 @@ export const CreateDatasetItem2DatasetsFile$outboundSchema: z.ZodType<
   CreateDatasetItem2DatasetsFile
 > = z.object({
   fileData: z.string(),
-  filename: z.string().optional(),
+  filename: z.string(),
 }).transform((v) => {
   return remap$(v, {
     fileData: "file_data",
@@ -5111,7 +5111,7 @@ export const CreateDatasetItemResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-06-17T09:09:46.815Z",
+    "2025-06-17T22:56:18.095Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -5175,7 +5175,7 @@ export const CreateDatasetItemResponseBody$outboundSchema: z.ZodType<
   createdById: z.string().optional(),
   updatedById: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-06-17T09:09:46.815Z"))
+  updated: z.date().default(() => new Date("2025-06-17T22:56:18.095Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {