+++ packages/orq-rc/src/models/operations/retrievedatapoint.ts
@@ -250,7 +250,7 @@ export type RetrieveDatapoint2File = {
   /**
    * The name of the file, used when passing the file to the model as a string.
    */
-  filename?: string | undefined;
+  filename: string;
 };
 
 export type RetrieveDatapoint24 = {
@@ -1459,7 +1459,7 @@ export const RetrieveDatapoint2File$inboundSchema: z.ZodType<
   unknown
 > = z.object({
   file_data: z.string(),
-  filename: z.string().optional(),
+  filename: z.string(),
 }).transform((v) => {
   return remap$(v, {
     "file_data": "fileData",
@@ -1469,7 +1469,7 @@ export const RetrieveDatapoint2File$inboundSchema: z.ZodType<
 /** @internal */
 export type RetrieveDatapoint2File$Outbound = {
   file_data: string;
-  filename?: string | undefined;
+  filename: string;
 };
 
 /** @internal */
@@ -1479,7 +1479,7 @@ export const RetrieveDatapoint2File$outboundSchema: z.ZodType<
   RetrieveDatapoint2File
 > = z.object({
   fileData: z.string(),
-  filename: z.string().optional(),
+  filename: z.string(),
 }).transform((v) => {
   return remap$(v, {
     fileData: "file_data",
@@ -2481,7 +2481,7 @@ export const RetrieveDatapointResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-06-17T09:09:46.815Z",
+    "2025-06-17T22:56:18.095Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -2539,7 +2539,7 @@ export const RetrieveDatapointResponseBody$outboundSchema: z.ZodType<
   createdById: z.string().optional(),
   updatedById: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-06-17T09:09:46.815Z"))
+  updated: z.date().default(() => new Date("2025-06-17T22:56:18.095Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {