+++ packages/orq-rc/src/models/operations/updatedatapoint.ts
@@ -240,7 +240,7 @@ export type UpdateDatapoint2File = {
   /**
    * The name of the file, used when passing the file to the model as a string.
    */
-  filename?: string | undefined;
+  filename: string;
 };
 
 export type UpdateDatapoint24 = {
@@ -712,7 +712,7 @@ export type UpdateDatapoint2DatasetsFile = {
   /**
    * The name of the file, used when passing the file to the model as a string.
    */
-  filename?: string | undefined;
+  filename: string;
 };
 
 export type UpdateDatapoint2Datasets4 = {
@@ -1833,7 +1833,7 @@ export const UpdateDatapoint2File$inboundSchema: z.ZodType<
   unknown
 > = z.object({
   file_data: z.string(),
-  filename: z.string().optional(),
+  filename: z.string(),
 }).transform((v) => {
   return remap$(v, {
     "file_data": "fileData",
@@ -1843,7 +1843,7 @@ export const UpdateDatapoint2File$inboundSchema: z.ZodType<
 /** @internal */
 export type UpdateDatapoint2File$Outbound = {
   file_data: string;
-  filename?: string | undefined;
+  filename: string;
 };
 
 /** @internal */
@@ -1853,7 +1853,7 @@ export const UpdateDatapoint2File$outboundSchema: z.ZodType<
   UpdateDatapoint2File
 > = z.object({
   fileData: z.string(),
-  filename: z.string().optional(),
+  filename: z.string(),
 }).transform((v) => {
   return remap$(v, {
     fileData: "file_data",
@@ -3962,7 +3962,7 @@ export const UpdateDatapoint2DatasetsFile$inboundSchema: z.ZodType<
   unknown
 > = z.object({
   file_data: z.string(),
-  filename: z.string().optional(),
+  filename: z.string(),
 }).transform((v) => {
   return remap$(v, {
     "file_data": "fileData",
@@ -3972,7 +3972,7 @@ export const UpdateDatapoint2DatasetsFile$inboundSchema: z.ZodType<
 /** @internal */
 export type UpdateDatapoint2DatasetsFile$Outbound = {
   file_data: string;
-  filename?: string | undefined;
+  filename: string;
 };
 
 /** @internal */
@@ -3982,7 +3982,7 @@ export const UpdateDatapoint2DatasetsFile$outboundSchema: z.ZodType<
   UpdateDatapoint2DatasetsFile
 > = z.object({
   fileData: z.string(),
-  filename: z.string().optional(),
+  filename: z.string(),
 }).transform((v) => {
   return remap$(v, {
     fileData: "file_data",
@@ -5046,7 +5046,7 @@ export const UpdateDatapointResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-06-17T09:09:46.815Z",
+    "2025-06-17T22:56:18.095Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -5108,7 +5108,7 @@ export const UpdateDatapointResponseBody$outboundSchema: z.ZodType<
   createdById: z.string().optional(),
   updatedById: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-06-17T09:09:46.815Z"))
+  updated: z.date().default(() => new Date("2025-06-17T22:56:18.095Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {